+++ docs/create.py
@@ -73,6 +73,12 @@ To check whether a file exists in the operating system:
 
    $ fuc fucexist example.txt
 
+To find all VCF files within the current directory:
+
+.. code-block:: console
+
+   $ fuc fucfind . vcf
+
 To count sequence reads in a FASTQ file:
 
 .. code-block:: console