+++ test.py
@@ -5,14 +5,15 @@ import subprocess
 from fuc.api.common import fuc_dir
 from fuc import pyvcf, pybed, pyfq
 
-vcf_file1 = f'{fuc_dir()}/data/vcf/1.vcf'
-vcf_file2 = f'{fuc_dir()}/data/vcf/2.vcf'
-vcf_file3 = f'{fuc_dir()}/data/vcf/3.vcf'
-bed_file1 = f'{fuc_dir()}/data/bed/1.bed'
-bed_file2 = f'{fuc_dir()}/data/bed/2.bed'
-fq_file1 = f'{fuc_dir()}/data/fq/1.fastq'
-text_file1 = f'{fuc_dir()}/data/text/1.txt'
-text_file2 = f'{fuc_dir()}/data/text/2.txt'
+fuc_dir = fuc_dir()
+vcf_file1 = f'{fuc_dir}/data/vcf/1.vcf'
+vcf_file2 = f'{fuc_dir}/data/vcf/2.vcf'
+vcf_file3 = f'{fuc_dir}/data/vcf/3.vcf'
+bed_file1 = f'{fuc_dir}/data/bed/1.bed'
+bed_file2 = f'{fuc_dir}/data/bed/2.bed'
+fq_file1 = f'{fuc_dir}/data/fq/1.fastq'
+text_file1 = f'{fuc_dir}/data/text/1.txt'
+text_file2 = f'{fuc_dir}/data/text/2.txt'
 
 class TestPyvcf(unittest.TestCase):
 
@@ -27,7 +28,7 @@ class TestPyvcf(unittest.TestCase):
 
     def test_filter_bed(self):
         vf = pyvcf.read_file(vcf_file1)
-        bf = pybed.read_file(f'{fuc_dir()}/data/bed/1.bed')
+        bf = pybed.read_file(f'{fuc_dir}/data/bed/1.bed')
         vf = vf.filter_bed(bf)
         self.assertEqual(vf.df.shape, (3, 13))
 
@@ -54,16 +55,16 @@ class TestPyvcf(unittest.TestCase):
 class TestPybed(unittest.TestCase):
 
     def test_intersect(self):
-        bf1 = pybed.read_file(f'{fuc_dir()}/data/bed/1.bed')
-        bf2 = pybed.read_file(f'{fuc_dir()}/data/bed/2.bed')
-        bf3 = pybed.read_file(f'{fuc_dir()}/data/bed/3.bed')
+        bf1 = pybed.read_file(f'{fuc_dir}/data/bed/1.bed')
+        bf2 = pybed.read_file(f'{fuc_dir}/data/bed/2.bed')
+        bf3 = pybed.read_file(f'{fuc_dir}/data/bed/3.bed')
         bf4 = bf1.intersect(bf2)
         self.assertEqual(bf3.to_string(), bf4.to_string())
 
 class TestPyfq(unittest.TestCase):
 
     def test_shape(self):
-        qf = pyfq.read_file(f'{fuc_dir()}/data/fq/1.fastq')
+        qf = pyfq.read_file(f'{fuc_dir}/data/fq/1.fastq')
         self.assertEqual(qf.shape, (5, 4))
 
 class TestCli(unittest.TestCase):
@@ -90,6 +91,10 @@ class TestCli(unittest.TestCase):
     def test_fucexist(self):
         result = subprocess.run(['fuc', 'fucexist', vcf_file1], capture_output=True, text=True, check=True)
         self.assertTrue('True' in result.stdout)
+
+    def test_fucfind(self):
+        result = subprocess.run(['fuc', 'fucfind', fuc_dir, 'vcf'], capture_output=True, text=True, check=True)
+        self.assertTrue('1.vcf' in result.stdout)
 
     def test_qfcount(self):
         result = subprocess.run(['fuc', 'qfcount', fq_file1], capture_output=True, text=True, check=True)